requirements.txt
requests==2.26.0
python-dotenv==1.0.1
beautifulsoup4==4.10.0
google-api-python-client==2.29.0
